+++ wxcadm/__init__.py
@@ -1,5 +1,3 @@
-name = 'wxcadm'
-
 import logging
 
 from .wxcadm import Webex, RedSky, XSIEvents, XSICallQueue
@@ -8,7 +6,7 @@ from .exceptions import (
     PutError, XSIError, NotAllowed, CSDMError
 )
 
-from .wxcadm import decode_spark_id
+from .wxcadm import decode_spark_id, console_logging
 
 # Set up NullHandler for logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())